# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/auth/verifyOTP": {
      "post": {
        "operationId": "verifyOtp",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOTPDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Cats example",
    "description": "The cats API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "mobile": {
            "type": "string",
            "minLength": 10,
            "maxLength": 10
          },
          "password": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "mobile",
          "password"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "name": {
            "type": "string"
          },
          "mobile": {
            "type": "string",
            "minLength": 10,
            "maxLength": 10
          },
          "userType": {
            "enum": [
              "client",
              "freelancer",
              "admin"
            ],
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "mobile",
          "userType"
        ]
      },
      "VerifyOTPDto": {
        "type": "object",
        "properties": {
          "mobile": {
            "type": "string",
            "minLength": 10,
            "maxLength": 10
          },
          "otp": {
            "type": "string"
          },
          "tokenId": {
            "type": "string"
          }
        },
        "required": [
          "mobile",
          "otp",
          "tokenId"
        ]
      },
...
    }
  }
}